+++ crates/remote-prover-client/src/remote_prover/batch_prover.rs
@@ -71,8 +71,8 @@ impl RemoteBatchProver {
 
         #[cfg(target_arch = "wasm32")]
         let new_client = {
-            let mut fetch_options =
-                tonic_web_wasm_client::FetchOptions::new().timeout(self.timeout);
+            let fetch_options =
+                tonic_web_wasm_client::options::FetchOptions::new().timeout(self.timeout);
             let web_client = tonic_web_wasm_client::Client::new_with_options(
                 self.endpoint.clone(),
                 fetch_options,